sched/core: Skip rq lock in try_to_wake_up() when WALT is disabled
When WALT is disabled, do_pl_notif() always returns false, in which case
this bit of code serves no purpose. As a result, this WALT-specific code
spins on acquiring the rq lock in a hot path, wasting CPU time. Compile
it out when WALT is disabled to eliminate the unnecessary overhead.

Change-Id: I6690741f54edaf3c63287e383fcbe7393ef69878
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Angga <angga@linuxmail.org>